Update error handling in LsmTree.
throwing exceptions at the API barrier for now. may switch to
std::expected later and since i'm using it everywhere else.

+++ src/Constants.h
@@ -2,6 +2,6 @@
 #include <cstddef>
 namespace lsm_storage_engine {
 namespace lsm_constants {
-constexpr size_t kMemTableFlushThreshold = 1UZ << 8;
+constexpr unsigned long kMemTableFlushThreshold = 1UZ << 8;
 }
 }; // namespace lsm_storage_engine
